+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Dohmann 2025 piperacillin ([doi:10.1002/bcp.70153](https://doi.org/10.1002/bcp.70153)) -- adults with renal impairment or on intermittent haemodialysis in a non-intensive-care ward setting.
+- Add Sun 2025 paclitaxel ([doi:10.1111/cts.70404](https://doi.org/10.1111/cts.70404)) -- adult women with breast cancer receiving weekly paclitaxel.
 
 - Add Siebinga 2023 [68Ga]Ga- and [177Lu]Lu-HA-DOTATATE ([doi:10.1186/s40658-023-00565-4](https://doi.org/10.1186/s40658-023-00565-4)) - adults with neuroendocrine tumors.